+++ DAO/tournoiDAO.php
@@ -274,6 +274,195 @@ class TournoiDAO
         return $req;
     }
 
+    public function closeTournois($id)
+    {
+        $req = $this->linkpdo->prepare('UPDATE tournoi SET estFerme = 1 WHERE Id_Tournoi = :id');
+        $req->execute(
+            array(
+                'id' => $id
+            )
+        );
+    }
+
+    public function getIDJeuxTournoi($idTournoi)
+    {
+        $req = $this->linkpdo->prepare('SELECT Id_Jeu FROM concerner where id_Tournoi = :id');
+        $req->execute(
+            array(
+                'id' => $idTournoi
+            )
+        );
+        $idsJeux = array(); //Tableau qui contiendra les id des jeux
+        while ($datas = $req->fetch()) {
+            array_push($idsJeux, $datas['Id_Jeu']);
+        }
+        return $idsJeux;
+    }
+
+    
+    public function addPoule($nom, $idTournoi, $idJeu)
+    {
+        $req = $this->linkpdo->prepare('INSERT INTO poule VALUES (NULL, :nom, :idTournoi, :idJeu)');
+        $req->execute(
+            array(
+                'nom' => $nom,
+                'idTournoi' => $idTournoi,
+                'idJeu' => $idJeu
+            )
+        );
+    }
+
+    
+    public function getIDPoule($idTournoi, $idJeu)
+    {
+        $req = $this->linkpdo->prepare('SELECT Id_Poule FROM poule WHERE id_Tournoi = :idTournoi AND id_jeu = :idJeu');
+        $req->execute(
+            array(
+                'idTournoi' => $idTournoi,
+                'idJeu' => $idJeu
+            )
+        );
+        $idsPoules = array(); //Tableau qui contiendra les id des poules
+        while ($datas = $req->fetch()) {
+            array_push($idsPoules, $datas['Id_Poule']);
+        }
+        return $idsPoules;
+    }
+
+        //Equipe inscrites sur un tournoi en fonction d'un jeu
+        public function getEquipeInscrites($idTournoi, $idJeu)
+        {
+            $req = $this->linkpdo->prepare('SELECT etre_inscrit.id_equipe FROM etre_inscrit, equipe WHERE id_Tournoi = :idTournoi AND etre_inscrit.id_Equipe = equipe.id_Equipe AND etre_inscrit.id_Jeu = :idJeu order by Nb_pts_Champ DESC');
+            $req->execute(
+                array(
+                    'idTournoi' => $idTournoi,
+                    'idJeu' => $idJeu
+                )
+            );
+            $idsEquipes = array(); //Tableau qui contiendra les id des poules
+            while ($datas = $req->fetch()) {
+                array_push($idsEquipes, $datas['id_equipe']);
+            }
+            return $idsEquipes;
+        }
+
+        
+    public function assignerPoule($idTournoi, $idPoule, $idEquipe)
+    {
+        $req = $this->linkpdo->prepare('UPDATE etre_inscrit SET id_poule = :idPoule WHERE id_Tournoi = :idTournoi AND id_Equipe = :idEquipe');
+        $req->execute(
+            array(
+                'idPoule' => $idPoule,
+                'idTournoi' => $idTournoi,
+                'idEquipe' => $idEquipe
+            )
+        );
+    }
+
+    public function addRencontre($idEquipe1, $idEquipe2, $idPoule)
+    {
+        $req = $this->linkpdo->prepare('INSERT INTO rencontre VALUES (NULL, :idEquipe1, :idEquipe2, :idPoule, NULL)');
+        $req->execute(
+            array(
+                'idEquipe1' => $idEquipe1,
+                'idEquipe2' => $idEquipe2,
+                'idPoule' => $idPoule
+            )
+        );
+    }
+    
+    public function getLastIDPoule()
+    {
+        $req = $this->linkpdo->prepare("SELECT MAX(Id_Poule) FROM poule");
+        $req->execute();
+        while ($datas = $req->fetch()) {
+            return $datas['MAX(Id_Poule)'];
+        }
+    }
+
+        //Fonction qui retourne le premier d'une poule a partir d'un tournoi et d'un jeu donnée 
+        public function getPremierPoule($idTournoi, $idJeu, $idPoule)
+        {
+            $req = $this->linkpdo->prepare('SELECT etre_inscrit.id_Equipe FROM etre_inscrit, poule WHERE poule.id_Tournoi = :idTournoi and poule.Id_Jeu = :idJeu and poule.id_Poule = etre_inscrit.id_poule and etre_inscrit.id_poule = :idPoule order by nb_Match_Gagne desc limit 1');
+            $req->execute(
+                array(
+                    'idTournoi' => $idTournoi,
+                    'idJeu' => $idJeu,
+                    'idPoule' => $idPoule
+                )
+            );
+            while ($row = $req->fetch()) {
+                return $row['id_Equipe'];
+            }
+        }
+
+            //Retourne le nombre de point total de la poule 
+    public function getNbPointPoule($idPoule)
+    {
+        $req = $this->linkpdo->prepare('SELECT sum(nb_Match_Gagne) as nbMatchJouer FROM etre_inscrit WHERE etre_inscrit.id_poule = :idPoule');
+        $req->execute(
+            array(
+                'idPoule' => $idPoule
+            )
+        );
+        while ($row = $req->fetch()) {
+            return $row['nbMatchJouer'];
+        }
+    }
+
+    
+    public function getPerdantFinale($idPoule)
+    {
+        $req = $this->linkpdo->prepare('select distinct id_Equipe from rencontre where Id_Poule = :idPoule and id_Equipe not in ( SELECT id_Equipe FROM rencontre WHERE id_Poule = :idPoule group by gagnant) union select DISTINCT Id_Equipe_1 from rencontre where Id_Poule = :idPoule and Id_Equipe_1 not in ( SELECT id_equipe FROM rencontre WHERE id_Poule = :idPoule group by gagnant);');
+        $req->execute(
+            array(
+                'idPoule' => $idPoule
+
+            )
+        );
+        $equipePouleFinale = array();
+        while ($row = $req->fetch()) {
+            array_push($equipePouleFinale, $row['id_Equipe']);
+        }
+        return $equipePouleFinale;
+    }
+
+    public function getResultatFinaux($idPoule)
+    {
+        $req = $this->linkpdo->prepare("SELECT equipe.id_equipe, count(gagnant), nom FROM rencontre, equipe WHERE rencontre.id_equipe = equipe.id_equipe AND id_Poule = :idPoule group by gagnant order by 2 desc");
+        $testReq = $req->execute(
+            array(
+                "idPoule" => $idPoule
+            )
+        );
+        return $req;
+    }
+
+    public function getMultiplicateur($idTournoi)
+    {
+        $req = $this->linkpdo->prepare('SELECT Nombre_point_max FROM tournoi WHERE id_Tournoi = :idTournoi');
+        $req->execute(
+            array(
+                'idTournoi' => $idTournoi
+            )
+        );
+
+        while ($row = $req->fetch()) {
+            return $row['Nombre_point_max'];
+        }
+    }
+
+    public function updateClassementEquipe($idEquipe, $nbPoint)
+    {
+        $req = $this->linkpdo->prepare('UPDATE equipe SET Nb_pts_Champ = Nb_pts_Champ + :nbPoint  WHERE id_Equipe = :idEquipe');
+        $req->execute(
+            array(
+                'idEquipe' => $idEquipe,
+                'nbPoint' => $nbPoint
+            )
+        );
+    }
+
