+++ controller/admin/modification-tournoi-controller.php
@@ -141,6 +141,45 @@ if (isset($_POST['ajouterJeu'])) {
     }
 }
 
+//Affichage de la liste de tout les jeux enregistrés dans la base de données
+function afficherListeJeuxTournoi($reqJeuduTournois, $reqNonPresentTurnois)
+{
+    $html = '';
+    while ($data = $reqJeuduTournois->fetch()) {
+
+        $html .= '<option selected value="' . $data['Id_Jeu'] . '">' . $data['Libelle'] . '</option>';
+    }
+    while ($data = $reqNonPresentTurnois->fetch()) {
+
+        $html .= '<option value="' . $data['Id_Jeu'] . '">' . $data['Libelle'] . '</option>';
+    }
+
+    return $html;
+}
+
+function afficherListeTypeTournoi($type){   
+    $html = '';                           
+if ($type == "Local") {
+    $html .=  '<option value="Local" selected>Local (100 points max)</option>';
+}else{
+    $html .= '<option value="Local">Local (100 points max)</option>';
+}
+if ($type == "National") {
+    $html .= '<option value="National" selected>National (200 points max)</option>';
+}else{
+    $html .= '<option value="National">National (200 points max)</option>';
+}
+if ($type == "International") {
+    $html .= '<option value="International" selected>International (300 points max)</option>';
+}else{
+    $html .= '<option value="International">International (300 points max)</option>';
+}
+
+return $html;
+}
+
+
+
 ob_start();
 require_once(realpath(dirname(__FILE__) . '/../../view/admin/modification-tournoi-view.html'));
 $buffer = ob_get_clean();
