+++ controller/visiteur/header-controller.php
@@ -18,6 +18,7 @@ switch ($nomPage) {
     case "enregistrer-ecurie-view.php":
     case "creation-tournoi-controller.php":
     case "modification-tournoi-controller.php":
+    case "fermer-tournoi-controller.php":
    
         if ($_SESSION['role'] == "gestionnaire") {
             echo $header->customizeHeader($_SESSION['role']);
@@ -25,7 +26,6 @@ switch ($nomPage) {
             header("Location: ../../index.php");
         }
         break;
-
 }
