+++ model/Tournoi.php
@@ -5,15 +5,7 @@
 
     // ==== Creation Tournoi ===
     //Fonction pour afficher les jeux sous forme de liste déroulante
-    function afficherListeJeux($reqJeu){
-        $html = '';
-        while ($data = $reqJeu->fetch()) {
 
-            $html .= '<option value="' . $data['Id_Jeu'] . '">' . $data['Libelle'] . '</option>';
-        }
-
-        return $html;
-    }
 
     function trierPar(string $by){
         $triTournois = new TriTournois();
@@ -38,44 +30,8 @@
         
     }
 
-    // ==== Modification Tournoi ===
     
-    //Affichage de la liste de tout les jeux enregistrés dans la base de données
-    function afficherListeJeuxTournoi($reqJeuduTournois, $reqNonPresentTurnois){
-    $html = '';
-    while ($data = $reqJeuduTournois->fetch()) {
 
-        $html .= '<option selected value="' . $data['Id_Jeu'] . '">' . $data['Libelle'] . '</option>';
-    }
-    while ($data = $reqNonPresentTurnois->fetch()) {
-
-        $html .= '<option value="' . $data['Id_Jeu'] . '">' . $data['Libelle'] . '</option>';
-    }
-
-    return $html;
-    }
-
-    //
-    function afficherListeTypeTournoi($type){   
-        $html = '';                           
-    if ($type == "Local") {
-        $html .=  '<option value="Local" selected>Local (100 points max)</option>';
-    }else{
-        $html .= '<option value="Local">Local (100 points max)</option>';
-    }
-    if ($type == "National") {
-        $html .= '<option value="National" selected>National (200 points max)</option>';
-    }else{
-        $html .= '<option value="National">National (200 points max)</option>';
-    }
-    if ($type == "International") {
-        $html .= '<option value="International" selected>International (300 points max)</option>';
-    }else{
-        $html .= '<option value="International">International (300 points max)</option>';
-    }
-
-    return $html;
-    }
 
     // DAO // 
 
@@ -168,6 +124,11 @@
     function tournoiId($id){
         $tournoiDAO = new TournoiDAO();
         return $tournoiDAO->tournoiId($id);
+    }
+
+    function tournoiIdNom($idTournoi){
+        $tournoiDAO = new TournoiDAO();
+        return $tournoiDAO->tournoiId($idTournoi)->fetch()['Nom'];
     }
